--- task8-p2p-chat/p2chat-client/src/main/kotlin/ru/mir/spbau/sd/chat/client/msg/UsersEventsReactor.kt
@@ -1,9 +1,0 @@
-package ru.mir.spbau.sd.chat.client.msg
-
-/**
- * Created by: Egor Gorbunov
- * Date: 2/6/17
- * Email: egor-mailbox@ya.com
- */
-interface UsersEventsReactor {
-}
