+++ task8-p2p-chat/p2chat-client/src/main/kotlin/ru/mir/spbau/sd/chat/client/msg/UsersMsgDispatcher.kt
@@ -1,0 +1,7 @@
+package ru.mir.spbau.sd.chat.client.msg
+
+/**
+ * Class,
+ */
+class UsersMsgDispatcher {
+}
