+++ task8-p2p-chat/p2chat-client/src/main/kotlin/ru/mir/spbau/sd/chat/client/msg/UsersNetEventsReactor.kt
@@ -5,5 +5,5 @@ package ru.mir.spbau.sd.chat.client.msg
  * Date: 2/6/17
  * Email: egor-mailbox@ya.com
  */
-interface ChatServerMsgListener {
+interface UsersNetEventsReactor {
 }
