+++ task8-p2p-chat/p2chat-client/src/main/kotlin/ru/mir/spbau/sd/chat/client/net/ClientChatNetworkController.kt
@@ -1,9 +1,9 @@
-package ru.mir.spbau.sd.chat.client.model
+package ru.mir.spbau.sd.chat.client.net
 
 /**
  * Created by: Egor Gorbunov
  * Date: 2/6/17
  * Email: egor-mailbox@ya.com
  */
-class OneUserMsgBox {
+class ClientChatNetworkController {
 }
