+++ task8-p2p-chat/p2chat-commons/build.gradle
@@ -22,6 +22,10 @@ repositories {
 
 dependencies {
     compile project(':p2chat-proto')
+    compile 'com.google.guava:guava:21.0'
+    compile group: 'org.slf4j', name: 'slf4j-api', version: '1.7.22'
+    compile group: 'ch.qos.logback', name: 'logback-core', version: '1.1.9'
+    compile group: 'ch.qos.logback', name: 'logback-classic', version: '1.1.9'
     compile "org.jetbrains.kotlin:kotlin-stdlib:$kotlin_version"
     testCompile group: 'junit', name: 'junit', version: '4.11'
 }
