+++ task8-p2p-chat/p2chat-commons/src/main/kotlin/ru/mit/spbau/sd/chat/commons/net/simpleProtocol.kt
@@ -2,6 +2,7 @@ package ru.mit.spbau.sd.chat.commons.net
 
 import ru.mit.spbau.sd.chat.commons.intSizeInBytes
 import ru.mit.spbau.sd.chat.commons.net.*
+import ru.mit.spbau.sd.chat.commons.net.state.*
 import java.nio.ByteBuffer
 
 /**
