+++ task8-p2p-chat/p2chat-commons/src/main/kotlin/ru/mit/spbau/sd/chat/commons/net/state/MessageRead.kt
@@ -1,4 +1,4 @@
-package ru.mit.spbau.sd.chat.commons.net
+package ru.mit.spbau.sd.chat.commons.net.state
 
 import java.nio.ByteBuffer
