+++ task8-p2p-chat/p2chat-commons/src/main/kotlin/ru/mit/spbau/sd/chat/commons/net/state/NothingToWrite.kt
@@ -1,10 +1,10 @@
-package ru.mit.spbau.sd.chat.commons.net
+package ru.mit.spbau.sd.chat.commons.net.state
 
 import java.nio.ByteBuffer
 
 /**
  * State represents, that there is no data to
- * write
+ * write. That is somewhat dummy writing state.
  */
 class NothingToWrite : WritingState {
     private val buf = ByteBuffer.allocate(0)
