+++ task8-p2p-chat/p2chat-proto/src/main/proto/message.proto
@@ -65,5 +65,4 @@ message PeerToPeerMsg {
     ChatMessage message = 1;
     ChatUserIpAddr userFromId = 2;
     ChatUserInfo userInfo = 3;
-
 }
